+++ backend/bid/src/main/java/com/dokidoki/bid/kafka/dto/KafkaAuctionEndDTO.java
@@ -16,12 +16,12 @@ import java.time.ZoneId;
 @ToString
 public class KafkaAuctionEndDTO {
 
-    private long auctionId;
-    private long sellerId;
-    private long buyerId;
-    private int finalPrice;
-    private int priceSize;
-    private long productId;
+    private Long auctionId;
+    private Long sellerId;
+    private Long buyerId;
+    private Long finalPrice;
+    private Long priceSize;
+    private Long productId;
     private String productName;
     private LocalDateTime endTime; // 끝나는 시점에 생성
 
@@ -38,7 +38,7 @@ public class KafkaAuctionEndDTO {
                 .build();
         return dto;
     }
-    public static KafkaAuctionEndDTO of(AuctionRealtime auctionRealtime, long buyerId) {
+    public static KafkaAuctionEndDTO of(AuctionRealtime auctionRealtime, Long buyerId) {
         KafkaAuctionEndDTO dto = KafkaAuctionEndDTO.builder()
                 .auctionId(auctionRealtime.getAuctionId())
                 .sellerId(auctionRealtime.getSellerId())
@@ -55,7 +55,7 @@ public class KafkaAuctionEndDTO {
     public KafkaAuctionEndDTO() {}
 
     @Builder
-    public KafkaAuctionEndDTO(long auctionId, long sellerId, long buyerId, int finalPrice, int priceSize, long productId, String productName, LocalDateTime endTime) {
+    public KafkaAuctionEndDTO(Long auctionId, Long sellerId, Long buyerId, Long finalPrice, Long priceSize, Long productId, String productName, LocalDateTime endTime) {
         this.auctionId = auctionId;
         this.sellerId = sellerId;
         this.buyerId = buyerId;
