+++ backend/bid/src/main/java/com/dokidoki/bid/kafka/dto/KafkaBidDTO.java
@@ -13,17 +13,17 @@ import java.time.LocalDateTime;
 @Builder
 public class KafkaBidDTO {
 
-    private long beforeWinnerId;
-    private long memberId;
-    private long auctionId;
+    private Long beforeWinnerId;
+    private Long memberId;
+    private Long auctionId;
     private String name;
-    private int highestPrice;
+    private Long highestPrice;
     private String productName;
-    private long productId;
+    private Long productId;
     private LocalDateTime bidTime;
     public KafkaBidDTO() {}
 
-    public static KafkaBidDTO of(AuctionRealtime auctionRealtime, AuctionBidReq req, LeaderBoardMemberResp resp, long memberId, long beforeWinnerId) {
+    public static KafkaBidDTO of(AuctionRealtime auctionRealtime, AuctionBidReq req, LeaderBoardMemberResp resp, Long memberId, Long beforeWinnerId) {
         KafkaBidDTO dto = KafkaBidDTO.builder()
                 .beforeWinnerId(beforeWinnerId)
                 .memberId(memberId)
@@ -37,7 +37,7 @@ public class KafkaBidDTO {
         return dto;
     }
 
-    public KafkaBidDTO(long beforeWinnerId, long memberId, long auctionId, String name, int highestPrice, String productName, long productId, LocalDateTime bidTime) {
+    public KafkaBidDTO(Long beforeWinnerId, Long memberId, Long auctionId, String name, Long highestPrice, String productName, Long productId, LocalDateTime bidTime) {
         this.beforeWinnerId = beforeWinnerId;
         this.memberId = memberId;
         this.auctionId = auctionId;
